receiptsense/requirements.txt
boto3==1.34.116
python-dateutil==2.9.0